Merged revision(s) 8563 from sandbox/adsk/3.0m: [https://trac.osgeo.org/mapguide/ticket/2547]: Attribute values are wrong in Selection Panel if feature source is joined. http://trac.osgeo.org/mapguide/ticket/2547, submit on behalf of Andy Zhang.
In selection panel, the order of attribute names is from 'LayerMetaData', which is same as the order of layer's property mappings. But the order of values is same as the order of properties in class definition. They are same in most cases. However, they may be different if feature source is joined or a view.

Now we use the order of properties in class definition for both cases. 

........


git-svn-id: http://svn.osgeo.org/mapguide/trunk@8616 a442ad82-b40c-0410-ae44-e6d7bd6e4a36