Merged revision(s) 8584 from sandbox/adsk/3.0m: [https://trac.osgeo.org/mapguide/ticket/2551],Zoom in/out raster file (sid file) in flexible web layout results in server down. http://trac.osgeo.org/mapguide/ticket/2551, submit on behalf of Andy Zhang.
For raster data sources, the property exists only when the data model type is FdoRasterDataModelType_Palette. But we call the GetProperty(L"Palette") when bits of pixel is 8. So for some raster files, FdoCommandException will be thrown several times, which results in Server unstable. Now we add a check of the data model type to fix the issue. 
........


git-svn-id: http://svn.osgeo.org/mapguide/trunk@8617 a442ad82-b40c-0410-ae44-e6d7bd6e4a36