Merged revision(s) 8616-8618 from trunk/MgDev: Merged revision(s) 8563 from sandbox/adsk/3.0m: [https://trac.osgeo.org/mapguide/ticket/2547]: Attribute values are wrong in Selection Panel if feature source is joined. http://trac.osgeo.org/mapguide/ticket/2547, submit on behalf of Andy Zhang.
In selection panel, the order of attribute names is from 'LayerMetaData', which is same as the order of layer's property mappings. But the order of values is same as the order of properties in class definition. They are same in most cases. However, they may be different if feature source is joined or a view.

Now we use the order of properties in class definition for both cases. 

........

........
Merged revision(s) 8584 from sandbox/adsk/3.0m:
[https://trac.osgeo.org/mapguide/ticket/2551],Zoom in/out raster file (sid file) in flexible web layout results in server down.
http://trac.osgeo.org/mapguide/ticket/2551, submit on behalf of Andy Zhang.

For raster data sources, the property exists only when the data model type is FdoRasterDataModelType_Palette. But we call the GetProperty(L"Palette") when bits of pixel is 8. So for some raster files, FdoCommandException will be thrown several times, which results in Server unstable. Now we add a check of the data model type to fix the issue. 
........

........
Merged revision(s) 8607 from sandbox/adsk/3.0m:
Fix Ticket [https://trac.osgeo.org/mapguide/ticket/2552]: Zoom slider not available in IE 10 with basic layout and without compatibility mode.

Submit on behalf of Andy Su.

The root case of the defect is that the feature DirectX-based Filters(filter:progid:DXImageTransform.Microsoft.AlphaImageLoader) has been removed from IE10’s Standards and Quirks modes. To fix the issue, use a standards-based alternative tag <img> to replace it. 
........

........


git-svn-id: http://svn.osgeo.org/mapguide/branches/2.6@8620 a442ad82-b40c-0410-ae44-e6d7bd6e4a36