Merged revision(s) 8700 from trunk/MgDev: Merged revision(s) 8694-8695 from sandbox/adsk/3.0m: [https://trac.osgeo.org/mapguide/ticket/2562], Linux: Select feature doesn't work when type of key column is string.
In Linux, when the key column of a feature class is string type, select feature on this layer doesn't work. The feature is not highlighted, the selection panel is empty either. 
I found that the API SelectFeatures works well. The value of the key column is correct. But after encoding it to base64 string, it becomes 'AAAA', which means an empty string. So the error is in encoding key. 
wchar is UTF32 in Linux, while it is UTF16 in Windows. In Linux, we will first convert the wide string to UTF16, then convert the UTF16 string to UTF8. Something is wrong during the 2 conversions. Now I change it to use UTF32toUTF8 directly. The result is correct after the change.
 
........
[https://trac.osgeo.org/mapguide/ticket/2562], Linux: Select feature doesn't work when type of key column is string.

Select feature doesn't work again when connecting to Oracle. The type of key column is NUMBER(10). And the FDO data type of the column is double. The odd thing is that it works well with my local build version, but don't work in AIMS official build. After investigation, I found the issue is in KeyEncode::WriteDouble(). The WriteDouble function of KeyEncode is like below:
void KeyEncode::WriteDouble(double d)
{
    FdoInt64 swap = *(FdoInt64*)&d;
    WriteInt64(swap);
}
The code  FdoInt64 swap = *(FdoInt64*)&d; is optimized by gcc. It works after I added the volatile keyword.
........

........


git-svn-id: http://svn.osgeo.org/mapguide/branches/3.0@8701 a442ad82-b40c-0410-ae44-e6d7bd6e4a36