Merged revision(s) 8871, 8912, 8921-8922, 8942 from sandbox/adsk/3.1n: Fix [https://trac.osgeo.org/mapguide/ticket/2585]: the total entries is not correct with Oracle data source.
It is because the initial vale of total entries is -1. We should set it to 0.
........
Fix the issue that DateTime value cannot be displayed in selection panel.
........
In some special FDO provider such as Autodesk RealDWG provider, the 'Geometry' property is already in property mappings. In method MgHtmlController::WriteSelectedFeatureAttributes(), we will first add all mapping properties, then add geometry property. We need to check if 'Geometry' is added to avoid adding it twice.
........
Some providers such as Sqlite provider, CreateCommand(FdoCommandType_Select) will return a extended select command. So we need to check if it is really a extended select command before using it. 
........
Fix the defect that zoom raster image results in server crash.
It is because if the displayed size of image is very small, the width or height may be 0. And we will use the width and height as the parameters of RESAMPLE. FDO Raster provider will crash in this case. I set width or height to 1 if its value is 0 to avoid the error.
........


git-svn-id: http://svn.osgeo.org/mapguide/branches/3.0@8959 a442ad82-b40c-0410-ae44-e6d7bd6e4a36