[https://trac.osgeo.org/mapguide/ticket/2732]: Use explicit property lists where MgFeatureQueryOptions is used.
This is to ensure that a "select * from featureclass" query from relational FDO providers will have its column list bounded to what is recognized in the class definition returned by the provider. Some providers are known to produce "select * from table" queries that "leak" out column types that the provider does not know how to process. Using an explicit property list from its class definition will allow us to avoid such issues.

This is for the Java implementation of the AJAX viewer

git-svn-id: http://svn.osgeo.org/mapguide/trunk@9004 a442ad82-b40c-0410-ae44-e6d7bd6e4a36