Merged revision(s) 9017 from sandbox/adsk/3.0m/Common/CoordinateSystem: [https://trac.osgeo.org/mapguide/ticket/2735]: Cannot read some user defined coordinate systems.
The CsMap dictionary string encoding was changed to UTF8 (see ticket [https://trac.osgeo.org/mapguide/ticket/2508] ​http://trac.osgeo.org/mapguide/ticket/2508). All strings in CsMap dictionaries were changed to UTF8 too. But if it is an user defined CS and the description of the CS contains characters which are not UTF8, loading the CS will fail. In order to load the CS, we remove invalid characters when loading it.



git-svn-id: http://svn.osgeo.org/mapguide/trunk@9018 a442ad82-b40c-0410-ae44-e6d7bd6e4a36