Merged revision(s) 9018-9021 from trunk/MgDev: Merged revision(s) 9017 from sandbox/adsk/3.0m/Common/CoordinateSystem: [https://trac.osgeo.org/mapguide/ticket/2735]: Cannot read some user defined coordinate systems.
The CsMap dictionary string encoding was changed to UTF8 (see ticket [https://trac.osgeo.org/mapguide/ticket/2508] ​http://trac.osgeo.org/mapguide/ticket/2508). All strings in CsMap dictionaries were changed to UTF8 too. But if it is an user defined CS and the description of the CS contains characters which are not UTF8, loading the CS will fail. In order to load the CS, we remove invalid characters when loading it.


........
[https://trac.osgeo.org/mapguide/ticket/2736]: Server crash when layer contains complicate symbols https://trac.osgeo.org/mapguide/ticket/2736

It is a heap corruption defect. The issue is at line 1792~1814, Common\Renderers\AGGRenderer.cpp. There are minx and maxx which are double. Then we convert them to int. But in some special cases, the converted imaxx will be less than iminx. It results in some memory are corrupted in next operations. To fix the issue, we need to add a check for the values.
........
[https://trac.osgeo.org/mapguide/ticket/2736]: Server crash when layer contains complicate symbols https://trac.osgeo.org/mapguide/ticket/2736

Refine submission 9019.  
........
Merged revision(s) 9019-9020 from sandbox/adsk/3.1n:
[https://trac.osgeo.org/mapguide/ticket/2736]: Server crash when layer contains complicate symbols https://trac.osgeo.org/mapguide/ticket/2736

It is a heap corruption defect. The issue is at line 1792~1814, Common\Renderers\AGGRenderer.cpp. There are minx and maxx which are double. Then we convert them to int. But in some special cases, the converted imaxx will be less than iminx. It results in some memory are corrupted in next operations. To fix the issue, we need to add a check for the values.
........
[https://trac.osgeo.org/mapguide/ticket/2736]: Server crash when layer contains complicate symbols https://trac.osgeo.org/mapguide/ticket/2736

Refine submission 9019.  
........

........


git-svn-id: http://svn.osgeo.org/mapguide/branches/3.0@9023 a442ad82-b40c-0410-ae44-e6d7bd6e4a36