Merged revision(s) 9021 from trunk/MgDev: Merged revision(s) 9019-9020 from sandbox/adsk/3.1n: [https://trac.osgeo.org/mapguide/ticket/2736]: Server crash when layer contains complicate symbols https://trac.osgeo.org/mapguide/ticket/2736
It is a heap corruption defect. The issue is at line 1792~1814, Common\Renderers\AGGRenderer.cpp. There are minx and maxx which are double. Then we convert them to int. But in some special cases, the converted imaxx will be less than iminx. It results in some memory are corrupted in next operations. To fix the issue, we need to add a check for the values.
........
[https://trac.osgeo.org/mapguide/ticket/2736]: Server crash when layer contains complicate symbols https://trac.osgeo.org/mapguide/ticket/2736

Refine submission 9019.  
........

........


git-svn-id: http://svn.osgeo.org/mapguide/branches/2.6@9024 a442ad82-b40c-0410-ae44-e6d7bd6e4a36