[https://trac.osgeo.org/mapguide/ticket/2754]: Fail to stylize layer when expression in symbol definition has invalid value
Some settings of a symbol definition e.g. width/height can be expressions. Suppose a table has a column HEIGHT which type is double, and the height of symbol is 0.1*HEIGHT. Then if a feature's HEIGHT is null, a FDOException will be thrown when stylizing the feature, and stylize layer will fail.

To fix this bug, we catch FDOException when applying styles.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@9106 a442ad82-b40c-0410-ae44-e6d7bd6e4a36