Merge fix for [https://trac.osgeo.org/mapguide/ticket/2750] to main branch.
[https://trac.osgeo.org/mapguide/ticket/2750]: Isapi mapagent crashes when server error message is too long

If there is an exception when MapGuide server processes a request, the  exception will be sent back to web extension. Isapi response handler will  convert this exception to html format. 
The crash happens when the error message is too long. We will get a  'security check fail or buffer overflow' error. We think it is related to  the unsafe API vsprintf(). It doesn't crash after changing the API to vsprintf_s().

git-svn-id: http://svn.osgeo.org/mapguide/trunk@9107 a442ad82-b40c-0410-ae44-e6d7bd6e4a36