Implement RFC 157
This adds a new overload of MgMap::Create that accepts:

 - map definition id
 - map name
 - image width
 - image height
 - view center X
 - view center Y
 - view scale
 - dpi

With this overload, the created MgMap instance can be used on the following APIs statelessly without needing to open it first from a session repository (via a session id and map name pair)

 - MgRenderingService::QueryFeatures
 - MgRenderingService::QueryFeatureProperties
 - MgRenderingService::RenderMapLegend
 - MgRenderingService::RenderDynamicOverlay

With this change, there is enough APIs in place from which to build a stateless map viewer that can perform selections and does not require creating or interacting with a session repository or resources inside it.

As an addendum to the original RFC, the MgSelectionBase::SetMap method has also been promoted to EXTERNAL_API to allow selections from an MgFeatureInformation to be interrogated without throwing.

A server unit test has been added to verify these new MgMap::Create parameters stick on the map.

A series of .net unit tests have been added to exercise the aforementioned rendering service APIs using the statelessly-created MgMap.

Fixes [https://trac.osgeo.org/mapguide/ticket/2766]

git-svn-id: http://svn.osgeo.org/mapguide/trunk@9143 a442ad82-b40c-0410-ae44-e6d7bd6e4a36