Merged revision(s) 9106-9107, 9147 from trunk/MgDev: [https://trac.osgeo.org/mapguide/ticket/2754]: Fail to stylize layer when expression in symbol definition has invalid value
Some settings of a symbol definition e.g. width/height can be expressions. Suppose a table has a column HEIGHT which type is double, and the height of symbol is 0.1*HEIGHT. Then if a feature's HEIGHT is null, a FDOException will be thrown when stylizing the feature, and stylize layer will fail.

To fix this bug, we catch FDOException when applying styles.
........
Merge fix for [https://trac.osgeo.org/mapguide/ticket/2750] to main branch.

[https://trac.osgeo.org/mapguide/ticket/2750]: Isapi mapagent crashes when server error message is too long

If there is an exception when MapGuide server processes a request, the  exception will be sent back to web extension. Isapi response handler will  convert this exception to html format. 
The crash happens when the error message is too long. We will get a  'security check fail or buffer overflow' error. We think it is related to  the unsafe API vsprintf(). It doesn't crash after changing the API to vsprintf_s().
........
[https://trac.osgeo.org/mapguide/ticket/2767]: Add missing JSON format conversion support for the following mapagent operations:

 - DESCRIBEDRAWING
 - ENUMERATEDRAWINGLAYERS
 - ENUMERATEDRAWINGSECTIONRESOURCES
 - ENUMERATEDRAWINGSECTIONS
 - ENUMERATEGROUPS
 - ENUMERATEUSERS
 - ENUMERATEREPOSITORIES
 - GETFEATURESETENVELOPE
 - GETREPOSITORYCONTENT
 - GETREPOSITORYHEADER
 - GETSITEINFO
 - GETSITESTATUS
 - GETSITEVERSION
 - GETTILEPROVIDERS
........


git-svn-id: http://svn.osgeo.org/mapguide/branches/3.1@9346 a442ad82-b40c-0410-ae44-e6d7bd6e4a36