Merged revision(s) 9393-9395 from branches/3.1/MgDev: Introduce a new preprocessor symbol SWIG_PUBLIC_API
With this symbol, we #ifdef out and forward-declare as much of ACE and xerces types and headers in our MapGuide common headers. The only ACE header that is leaked out now is the header for ACE_Recursive_Thread_Mutex as that is present as member of several classes in the public MapGuide API surface.

In our SWIG binding projects, SWIG_PUBLIC_API is defined. All other consuming projects do not define this symbol so they see the original set of ACE and xerces headers.

This change slightly breaks the SWIG C++ glue code for .net (which used un-qualified list), so our internal copy of SWIG is updated to use fully qualified list (std::list) and our various SWIG parameter files include the <list> header.

This change eliminates the need for us to maintain a patched copy of TSRM/readdir.h for PHP, so that is removed.

Fixes [https://trac.osgeo.org/mapguide/ticket/2781]
........
[https://trac.osgeo.org/mapguide/ticket/2781]: Remove xerces include and linker settings
........
[https://trac.osgeo.org/mapguide/ticket/2781]: Limit the scope of these changes to Windows. Linux never had these conflicts, so no need to apply any of these changes there.
........


git-svn-id: http://svn.osgeo.org/mapguide/trunk@9396 a442ad82-b40c-0410-ae44-e6d7bd6e4a36