Merged revision(s) 9439-9440, 9442, 9447, 9449-9450 from sandbox/jng/deprecate_dwf: Comment out and remove usage of the DWF Toolkit impl of zip file reading and writing in MgServerResourceService
In its place, we'll use the minizip contrib library as part of the zlib sources. This commit sets up the projects to build minizip and for MgServerResourceService to link to it and zlib.
........
Initial implementation of minizip-based MGP reading and writing.
........
Remove ZipFileHandler as switching to minizip means there is no longer a common base for MgZipFileReader/MgZipFileWriter to inherit from.
........
Fix bad linker paths in Debug
........
Fix improper close call on zip reader/writer that caused the zip file handles to remain open, causing the package tests to fail. Throw exceptions where appropriate.
........
Fix up Linux build
........


git-svn-id: http://svn.osgeo.org/mapguide/trunk@9488 a442ad82-b40c-0410-ae44-e6d7bd6e4a36